Kill one more class
Just parse jvm arguments, if any, and invoker will
use it or not use it, imple dependant.
